organized join team page and job listing page, added server-side cache to CMS requests
